mds: fix frozen inode check in MDCache::handle_discover()
When MDCache::handle_discover() encounters a frozen dirfrag, it should
proceed if the dirfrag is being merged, but the MDS hasn't frozen all
dirfrags yet. So when MDCache::handle_discover() checks if a inode is
frozen, it should use CInode::is_frozen_inode() (which doesn't check if
parent dirfrag is frozen).

Signed-off-by: Yan, Zheng <zheng.z.yan@intel.com>